+++ clash/src/gui/lobby/mod.rs
@@ -1,3 +1,4 @@
+use std::borrow::Cow;
 use std::mem::ManuallyDrop;
 use std::rc::Rc;
 use std::thread::JoinHandle;
@@ -5,10 +6,7 @@ use std::thread::JoinHandle;
 use clash_lib::lobby::{GamePhase, NetworkedLobby};
 use clash_lib::net::{LobbyMessage, Message};
 use clash_lib::PlayerId;
-use eframe::egui::{
-    Align, Button, CentralPanel, Layout, Response, SidePanel, Ui, Widget, WidgetText,
-};
-use eframe::epaint::Color32;
+use eframe::egui::{Align, Button, CentralPanel, Layout, SidePanel, Ui};
 use eframe::App;
 use itertools::intersperse;
 
@@ -21,7 +19,7 @@ use tracker::Tracker;
 
 use super::main_menu::MainMenu;
 use super::val_text::ValText;
-use super::GuiReceiver;
+use super::{GuiReceiver, UiExt};
 
 mod player_ui;
 mod tracker;
@@ -128,7 +126,7 @@ impl App for Game {
                     }
                 }
                 GamePhase::Playing => {
-                    ui.add(Tracker::new(&self.lobby, self.local_player_id));
+                    Tracker::new(&self.state, &self.lobby, self.local_player_id).ui(ui);
                 }
                 GamePhase::Finished => todo!(),
             }
@@ -155,41 +153,32 @@ impl Game {
     }
 
     fn options_controls(&mut self, ui: &mut Ui) {
-        let mut updated_options = None;
+        let mut updated_options = Cow::Borrowed(&self.lobby.options);
 
-        let mut ng_plus = self.lobby.options.ng_plus;
-        ui.add_option("New Game+", &mut ng_plus, |&x| {
-            updated_options
-                .get_or_insert_with(|| self.lobby.options.clone())
-                .ng_plus = x;
+        ui.add_option("New Game+", updated_options.ng_plus, |x| {
+            updated_options.to_mut().ng_plus = x;
         })
         .on_hover_text(
             "All players start the game with the Bubble Bowl and Cruise Missile unlocked.",
         );
 
-        ui.add_option("Lab Door Cost", &mut self.lab_door_cost, |&n| {
-            updated_options
-                .get_or_insert_with(|| self.lobby.options.clone())
-                .lab_door_cost = n;
+        ui.add_option("Lab Door Cost", &mut self.lab_door_cost, |n| {
+            updated_options.to_mut().lab_door_cost = n;
         })
         .on_hover_text("Spatulas required to enter Chum Bucket Labs");
 
         ui.collapsing("Debug Options", |ui| {
-            ui.add_option("Tier Count", &mut self.tier_count, |&n| {
-                updated_options
-                    .get_or_insert_with(|| self.lobby.options.clone())
-                    .tier_count = n;
+            ui.add_option("Tier Count", &mut self.tier_count, |n| {
+                updated_options.to_mut().tier_count = n;
             })
-            .on_hover_text("Number of times a spatula can be collectd before it's disabled");
+            .on_hover_text("Number of times a spatula can be collected before it's disabled");
 
-            ui.add_option("Scores", self.scores.as_mut_slice(), |&(i, x)| {
-                updated_options
-                    .get_or_insert_with(|| self.lobby.options.clone())
-                    .spat_scores[i] = x;
+            ui.add_option("Scores", self.scores.as_mut_slice(), |(i, x)| {
+                updated_options.to_mut().spat_scores[i] = x;
             });
         });
 
-        if let Some(options) = updated_options {
+        if let Cow::Owned(options) = updated_options {
             self.lobby_data
                 .network_sender
                 .blocking_send(NetCommand::Send(Message::Lobby(
@@ -217,7 +206,11 @@ impl Game {
                 .on_disabled_hover_text(format!(
                     "Waiting on: {}",
                     intersperse(
-                        self.lobby.players.values().map(|p| p.options.name.as_str()),
+                        self.lobby
+                            .players
+                            .values()
+                            .filter(|p| !p.ready_to_start)
+                            .map(|p| p.options.name.as_str()),
                         ", "
                     )
                     .collect::<String>()
@@ -240,88 +233,5 @@ impl Game {
                 .try_send(NetCommand::Send(Message::Lobby(LobbyMessage::GameBegin {})))
                 .unwrap();
         }
-    }
-}
-
-trait UiExt<'a, In: ?Sized, Out> {
-    fn add_option(
-        &mut self,
-        label: impl Into<WidgetText>,
-        input: &'a mut In,
-        on_changed: impl FnMut(&Out) + 'a,
-    ) -> Response;
-}
-
-impl<'a, In, Out> UiExt<'a, In, Out> for Ui
-where
-    In: 'a + ?Sized,
-    Out: 'a,
-    OptionEditor<'a, In, Out>: Widget,
-{
-    fn add_option(
-        &mut self,
-        text: impl Into<WidgetText>,
-        input: &'a mut In,
-        on_changed: impl FnMut(&Out) + 'a,
-    ) -> Response {
-        let editor = OptionEditor::new(text, input, on_changed);
-        self.add(editor)
-    }
-}
-
-struct OptionEditor<'a, In: ?Sized, Out> {
-    label: WidgetText,
-    input: &'a mut In,
-    on_changed: Box<dyn FnMut(&Out) + 'a>,
-}
-
-impl<'a, In: ?Sized, Out> OptionEditor<'a, In, Out> {
-    fn new(text: impl Into<WidgetText>, input: &'a mut In, changed: impl FnMut(&Out) + 'a) -> Self {
-        Self {
-            label: text.into(),
-            input,
-            on_changed: Box::new(changed),
-        }
-    }
-}
-
-impl<'a, T> Widget for OptionEditor<'a, ValText<T>, T> {
-    fn ui(mut self, ui: &mut Ui) -> eframe::egui::Response {
-        ui.horizontal(|ui| {
-            if !self.input.is_valid() {
-                ui.style_mut().visuals.override_text_color = Some(Color32::DARK_RED);
-            }
-            ui.label(self.label);
-            if ui.text_edit_singleline(self.input).changed() && self.input.is_valid() {
-                if let Some(x) = self.input.get_val() {
-                    (self.on_changed)(x);
-                }
-            }
-        })
-        .response
-    }
-}
-
-impl<'a> Widget for OptionEditor<'a, bool, bool> {
-    fn ui(mut self, ui: &mut Ui) -> Response {
-        ui.horizontal(|ui| {
-            if ui.checkbox(self.input, self.label).changed() {
-                (self.on_changed)(self.input);
-            }
-        })
-        .response
-    }
-}
-
-impl<'a, T: Copy> Widget for OptionEditor<'a, [ValText<T>], (usize, T)> {
-    fn ui(mut self, ui: &mut Ui) -> Response {
-        ui.collapsing(self.label, |ui| {
-            for (i, input) in self.input.iter_mut().enumerate() {
-                ui.add_option(format!("Tier {}", i + 1), input, |&x| {
-                    (self.on_changed)(&(i, x))
-                });
-            }
-        })
-        .header_response
     }
 }